// src/app/interfaces/product.ts
export interface Product {
  description: string;
  groupID: number;
  nmfcsub: string;
  nmfc: string;
  hazMat: number;
  productID: number;
  active: number;
  auditStatus: number;
  auditor: number;
  densityMin: number;
  densityMax: number;
  productClass: number;
  productGroupID: number;
  code: string;
}